add establishment lead gateway routes and router
