+++ front/src/config/dependencies.ts
@@ -9,6 +9,7 @@ import { AssessmentGateway } from "src/core-logic/ports/AssessmentGateway";
 import { ConventionGateway } from "src/core-logic/ports/ConventionGateway";
 import { DeviceRepository } from "src/core-logic/ports/DeviceRepository";
 import { EstablishmentGateway } from "src/core-logic/ports/EstablishmentGateway";
+import { EstablishmentLeadGateway } from "src/core-logic/ports/EstablishmentLeadGateway";
 import { FormCompletionGateway } from "src/core-logic/ports/FormCompletionGateway";
 import { InclusionConnectedGateway } from "src/core-logic/ports/InclusionConnectedGateway";
 import { NavigationGateway } from "src/core-logic/ports/NavigationGateway";
@@ -24,6 +25,7 @@ export type Dependencies = {
   addressGateway: AddressGateway;
   technicalGateway: TechnicalGateway;
   establishmentGateway: EstablishmentGateway;
+  establishmentLeadGateway: EstablishmentLeadGateway;
   conventionGateway: ConventionGateway;
   searchGateway: SearchGateway;
   inclusionConnectedGateway: InclusionConnectedGateway;
