+++ front/src/core-logic/adapters/EstablishmentGateway/SimulatedEstablishmentGateway.ts
@@ -1,4 +1,4 @@
-import { Observable, delay, of, throwError } from "rxjs";
+import { Observable, delay, of } from "rxjs";
 import {
   BackOfficeJwt,
   ConventionJwt,
@@ -58,14 +58,5 @@ export class SimulatedEstablishmentGateway implements EstablishmentGateway {
         : currentEstablishment,
     );
     return of(undefined).pipe(delay(this.delay));
-  }
-
-  public rejectEstablishmentLeadRegistration$(
-    jwt: ConventionJwt,
-  ): Observable<void> {
-    console.log("rejectEstablishmentLeadRegistration", jwt);
-    return jwt === failedJwt
-      ? throwError(new Error("Failed Jwt"))
-      : of(undefined).pipe(delay(this.delay));
   }
 }
