+++ front/src/core-logic/adapters/EstablishmentGateway/TestEstablishmentGateway.ts
@@ -1,7 +1,6 @@
 import { Observable, Subject } from "rxjs";
 import {
   BackOfficeJwt,
-  ConventionJwt,
   EstablishmentJwt,
   FormEstablishmentDto,
   SiretDto,
@@ -18,8 +17,6 @@ export class TestEstablishmentGateway implements EstablishmentGateway {
   public establishmentModificationResponse$ = new Subject<void>();
 
   public formEstablishment$ = new Subject<FormEstablishmentDto>();
-
-  public unregisterEstablishmentLeadResponse$ = new Subject<void>();
 
   public addFormEstablishment$(
     _formEstablishment: FormEstablishmentDto,
@@ -50,11 +47,5 @@ export class TestEstablishmentGateway implements EstablishmentGateway {
     _jwt: EstablishmentJwt | BackOfficeJwt,
   ): Observable<void> {
     return this.editFormEstablishmentResult$;
-  }
-
-  public rejectEstablishmentLeadRegistration$(
-    _jwt: ConventionJwt,
-  ): Observable<void> {
-    return this.unregisterEstablishmentLeadResponse$;
   }
 }
