+++ front/src/core-logic/ports/EstablishmentLeadGateway.ts
@@ -1,0 +1,6 @@
+import { Observable } from "rxjs";
+import { ConventionJwt } from "shared";
+
+export interface EstablishmentLeadGateway {
+  rejectEstablishmentLeadRegistration$(jwt: ConventionJwt): Observable<void>;
+}
