+++ front/src/core-logic/storeConfig/createTestStore.ts
@@ -3,6 +3,7 @@ import { Dependencies } from "src/config/dependencies";
 import { TestAdminGateway } from "src/core-logic/adapters/AdminGateway/TestAdminGateway";
 import { InMemoryConventionGateway } from "src/core-logic/adapters/Convention/InMemoryConventionGateway";
 import { createTestDeviceRepository } from "src/core-logic/adapters/DeviceRepository/createTestDeviceRepository";
+import { TestEstablishmentLeadGateway } from "src/core-logic/adapters/EstablishmentLeadGateway/TestEstablishmentLeadGateway";
 import { TestInclusionConnectedGateway } from "src/core-logic/adapters/InclusionConnected/TestInclusionConnectedGateway";
 import { InMemoryNavigationGateway } from "src/core-logic/adapters/NavigationGateway/InMemoryNavigationGateway";
 import { TestTechnicalGateway } from "src/core-logic/adapters/TechnicalGateway/TestTechnicalGateway";
@@ -25,6 +26,7 @@ const createTestDependencies = () =>
     conventionGateway: new InMemoryConventionGateway(),
     deviceRepository: createTestDeviceRepository(),
     establishmentGateway: new TestEstablishmentGateway(),
+    establishmentLeadGateway: new TestEstablishmentLeadGateway(),
     formCompletionGateway: new TestFormCompletionGateway(),
     inclusionConnectedGateway: new TestInclusionConnectedGateway(),
     minSearchResultsToPreventRefetch: 2,
